+++ src/http/request.hpp
@@ -36,6 +36,7 @@ namespace http {
             bool isMIMEAccepted(const std::string&) const;
             bool isEncodingAccepted(const std::string&) const;
             inline bool isURIBad() const { return hasBadURI; };
+            inline bool getHasExplicitlyDefinedHTTPVersion0_9() const { return hasExplicitlyDefinedHTTPVersion0_9; };
 
             bool isFileValid(Response& response, const File& file) const;
             bool isInDocumentRoot(Response&, const std::string&) const;
@@ -55,6 +56,7 @@ namespace http {
             std::string pathStr;
             std::string rawPathStr; // The path BEFORE URI decoding
             bool hasBadURI = false; // Set to true if the decodeURI method fails, handled by Response object
+            bool hasExplicitlyDefinedHTTPVersion0_9; // Set to true if the status line has HTTP/0.9 explicitly in it (not allowed)
 
             std::string httpVersionStr;