+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     - Install via `sudo apt install php8.3-fpm`
 - Bump C++ version from C++17 to C++20
 - Add "Getting Started" section to README
+- Fix explicitly declaring HTTP/0.9 as the version in the status line (#108)
 
 ## v0.11.0
 - Now checks for a new version at startup (#91)